feature/packages_structure : the hilt dependencies were added, also the basic login screen was created with its viewModel and the packages structure of the presentation layer was set
