+++ app/src/main/java/com/chemasmas/fakestoreapi/FakeStoreApplication.kt
@@ -1,0 +1,8 @@
+package com.chemasmas.fakestoreapi
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class FakeStoreApplication : Application() {
+}
