+++ app/src/main/java/com/chemasmas/fakestoreapi/presentation/features/login/LoginViewModel.kt
@@ -1,0 +1,9 @@
+package com.chemasmas.fakestoreapi.presentation.features.login
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class LoginViewModel @Inject constructor() : ViewModel() {
+}
