+++ build.gradle
@@ -1,10 +1,12 @@
 buildscript {
     ext {
         compose_ui_version = '1.2.0'
+        compose_hilt_navigation_version = '1.0.0'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '7.4.2' apply false
     id 'com.android.library' version '7.4.2' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
+    id 'com.google.dagger.hilt.android' version '2.44' apply false
 }
